AsNoTracking/SAveChangesInterceptor

+++ Core/ECommerceAPI.Domain/Entities/Product.cs
@@ -11,7 +11,8 @@ namespace ECommerceAPI.Domain.Entities
     {
         public string Name { get; set; } = string.Empty;
         public int Stock { get; set; }
-        public long Price { get; set; }
+        public float Price { get; set; }
         public ICollection<Order>? Orders { get; set; }
+
     }
 }
